+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * `UserPolicyHandler` uses RolePermission repo and handles finding no roles
 * Value Object cleanup
 * Better migration warning on integration tests
+* Moved most `UserPolicyHandler` tests to unit tests
 
 ### Fixed